Enum bollard::errors::Error

pub enum Error {

    DockerResponseServerError {
        status_code: u16,
        message: String,
    },
    JsonDataError {
        message: String,
        column: usize,
    },
    APIVersionParseError {
        api_version: String,
    },
    RequestTimeoutError,
    DockerStreamError {
        error: String,
    },
    DockerContainerWaitError {
        error: String,
        code: i64,
    },
    MissingSessionBuildkitError {},
    JsonSerdeError {
        err: Error,
    },
    StrParseError {
        err: Utf8Error,
    },
    IOError {
        err: Error,
    },
    StrFmtError {
        err: Error,
    },
    HttpClientError {
        err: Error,
    },
    HyperResponseError {
        err: Error,
    },
    URLEncodedError {
        err: Error,
    },

}

The type of error embedded in an Error.

Variants

DockerResponseServerError

Fields

status_code: u16

Status code returned by the docker server.

message: String

Message returned by the docker server.

Generic error emitted by the docker server.

JsonDataError

Fields

message: String

Short section of the json close to the error.

column: usize

Character sequence at error location.

Error facilitating debugging failed JSON parsing.

APIVersionParseError

Fields

api_version: String

The api version returned by the server.

Error emitted when the docker is requested to build with buildkit without a session id

RequestTimeoutError

Error emitted when a request times out.

DockerStreamError

Fields

error: String

error string emitted by the Stream

Error emitted mid-stream as part of a successful docker operation

DockerContainerWaitError

Fields

error: String

error string returned from container wait call

code: i64

error code returned from container wait call

Error emitted as part of a container wait response

MissingSessionBuildkitError

Fields

Error emitted when a session is not provided to the buildkit engine

JsonSerdeError

Fields

err: Error

The original error emitted by serde.

Error emitted when JSON fails to serialize.

StrParseError

Fields

err: Utf8Error

The original error emitted.

Error emitted when log output generates an I/O error.

IOError

Fields

err: Error

The original error emitted.

Error emitted from an I/O error.

StrFmtError

Fields

err: Error

The original error emitted.

Error emitted from a formatting error.

HttpClientError

Fields

err: Error

The original error emitted.

Error emitted from an HTTP error.

HyperResponseError

Fields

err: Error

The original error emitted.

Error emitted from an HTTP error.

URLEncodedError

Fields

err: Error

The original error emitted.

Error emitted when serde fails to urlencod a struct of options