Struct bollard::image::BuildImageOptions

pub struct BuildImageOptions<T>where
    T: Into<String> + Eq + Hash + Serialize,{

    pub dockerfile: T,
    pub t: T,
    pub extrahosts: Option<T>,
    pub remote: T,
    pub q: bool,
    pub nocache: bool,
    pub cachefrom: Vec<T>,
    pub pull: bool,
    pub rm: bool,
    pub forcerm: bool,
    pub memory: Option<u64>,
    pub memswap: Option<i64>,
    pub cpushares: Option<u64>,
    pub cpusetcpus: T,
    pub cpuperiod: Option<u64>,
    pub cpuquota: Option<u64>,
    pub buildargs: HashMap<T, T>,
    pub shmsize: Option<u64>,
    pub squash: bool,
    pub labels: HashMap<T, T>,
    pub networkmode: T,
    pub platform: T,
    pub version: BuilderVersion,

}

Parameters to the Build Image API

Examples

use bollard::image::BuildImageOptions;

BuildImageOptions {
    dockerfile: "Dockerfile",
    t: "my-image",
    ..Default::default()
};

BuildImageOptions::<String> {
    ..Default::default()
};

Fields
dockerfile: T

Path within the build context to the Dockerfile. This is ignored if remote is specified and points to an external Dockerfile.

t: T

A name and optional tag to apply to the image in the name:tag format. If you omit the tag the default latest value is assumed. You can provide several t parameters.

extrahosts: Option<T>

Extra hosts to add to /etc/hosts.

remote: T

A Git repository URI or HTTP/HTTPS context URI. If the URI points to a single text file, the file’s contents are placed into a file called Dockerfile and the image is built from that file. If the URI points to a tarball, the file is downloaded by the daemon and the contents therein used as the context for the build. If the URI points to a tarball and the dockerfile parameter is also specified, there must be a file with the corresponding path inside the tarball.

q: bool

Suppress verbose build output.

nocache: bool

Do not use the cache when building the image.

cachefrom: Vec<T>

JSON array of images used for build cache resolution.

pull: bool

Attempt to pull the image even if an older image exists locally.

rm: bool

Remove intermediate containers after a successful build.

forcerm: bool

Always remove intermediate containers, even upon failure.

memory: Option<u64>

Set memory limit for build.

memswap: Option<i64>

Total memory (memory + swap). Set as -1 to disable swap.

cpushares: Option<u64>

CPU shares (relative weight).

cpusetcpus: T

CPUs in which to allow execution (e.g., 0-3, 0,1).

cpuperiod: Option<u64>

The length of a CPU period in microseconds.

cpuquota: Option<u64>

Microseconds of CPU time that the container can get in a CPU period.

buildargs: HashMap<T, T>

JSON map of string pairs for build-time variables. Users pass these values at build-time. Docker uses the buildargs as the environment context for commands run via the Dockerfile RUN instruction, or for variable expansion in other Dockerfile instructions.

shmsize: Option<u64>

Size of /dev/shm in bytes. The size must be greater than 0. If omitted the system uses 64MB.

squash: bool

Squash the resulting images layers into a single layer.

labels: HashMap<T, T>

Arbitrary key/value labels to set on the image, as a JSON map of string pairs.

networkmode: T

Sets the networking mode for the run commands during build. Supported standard values are: bridge, host, none, and container:<name|id>. Any other value is taken as a custom network’s name to which this container should connect to.

platform: T

Platform in the format os[/arch[/variant]]

version: BuilderVersion

Builder version to use