Struct bollard::models::NetworkCreateRequest

pub struct NetworkCreateRequest {
    pub name: String,
    pub check_duplicate: Option<bool>,
    pub driver: Option<String>,
    pub internal: Option<bool>,
    pub attachable: Option<bool>,
    pub ingress: Option<bool>,
    pub ipam: Option<Ipam>,
    pub enable_ipv6: Option<bool>,
    pub options: Option<HashMap<String, String, RandomState>>,
    pub labels: Option<HashMap<String, String, RandomState>>,
}

Fields
name: String

The network’s name.

check_duplicate: Option<bool>

Check for networks with duplicate names. Since Network is primarily keyed based on a random ID and not on the name, and network name is strictly a user-friendly alias to the network which is uniquely identified using ID, there is no guaranteed way to check for duplicates. CheckDuplicate is there to provide a best effort checking of any networks which has the same name but it is not guaranteed to catch all name collisions.

driver: Option<String>

Name of the network driver plugin to use.

internal: Option<bool>

Restrict external access to the network.

attachable: Option<bool>

Globally scoped network is manually attachable by regular containers from workers in swarm mode.

ingress: Option<bool>

Ingress network is the network which provides the routing-mesh in swarm mode.

ipam: Option<Ipam>

Optional custom IP scheme for the network.

enable_ipv6: Option<bool>

Enable IPv6 on the network.

options: Option<HashMap<String, String, RandomState>>

Network specific options to be used by the drivers.

labels: Option<HashMap<String, String, RandomState>>

User-defined key/value metadata.