Struct bollard::models::SystemInfo

pub struct SystemInfo {

    pub id: Option<String>,
    pub containers: Option<i64>,
    pub containers_running: Option<i64>,
    pub containers_paused: Option<i64>,
    pub containers_stopped: Option<i64>,
    pub images: Option<i64>,
    pub driver: Option<String>,
    pub driver_status: Option<Vec<Vec<String, Global>, Global>>,
    pub docker_root_dir: Option<String>,
    pub plugins: Option<PluginsInfo>,
    pub memory_limit: Option<bool>,
    pub swap_limit: Option<bool>,
    pub kernel_memory: Option<bool>,
    pub kernel_memory_tcp: Option<bool>,
    pub cpu_cfs_period: Option<bool>,
    pub cpu_cfs_quota: Option<bool>,
    pub cpu_shares: Option<bool>,
    pub cpu_set: Option<bool>,
    pub pids_limit: Option<bool>,
    pub oom_kill_disable: Option<bool>,
    pub ipv4_forwarding: Option<bool>,
    pub bridge_nf_iptables: Option<bool>,
    pub bridge_nf_ip6tables: Option<bool>,
    pub debug: Option<bool>,
    pub n_fd: Option<i64>,
    pub n_goroutines: Option<i64>,
    pub system_time: Option<String>,
    pub logging_driver: Option<String>,
    pub cgroup_driver: Option<SystemInfoCgroupDriverEnum>,
    pub cgroup_version: Option<SystemInfoCgroupVersionEnum>,
    pub n_events_listener: Option<i64>,
    pub kernel_version: Option<String>,
    pub operating_system: Option<String>,
    pub os_version: Option<String>,
    pub os_type: Option<String>,
    pub architecture: Option<String>,
    pub ncpu: Option<i64>,
    pub mem_total: Option<i64>,
    pub index_server_address: Option<String>,
    pub registry_config: Option<RegistryServiceConfig>,
    pub generic_resources: Option<GenericResourcesInner>,
    pub http_proxy: Option<String>,
    pub https_proxy: Option<String>,
    pub no_proxy: Option<String>,
    pub name: Option<String>,
    pub labels: Option<Vec<String, Global>>,
    pub experimental_build: Option<bool>,
    pub server_version: Option<String>,
    pub cluster_store: Option<String>,
    pub cluster_advertise: Option<String>,
    pub runtimes: Option<HashMap<String, Runtime, RandomState>>,
    pub default_runtime: Option<String>,
    pub swarm: Option<SwarmInfo>,
    pub live_restore_enabled: Option<bool>,
    pub isolation: Option<SystemInfoIsolationEnum>,
    pub init_binary: Option<String>,
    pub containerd_commit: Option<Commit>,
    pub runc_commit: Option<Commit>,
    pub init_commit: Option<Commit>,
    pub security_options: Option<Vec<String, Global>>,
    pub product_license: Option<String>,
    pub default_address_pools: Option<Vec<SystemInfoDefaultAddressPools, Global>>,
    pub warnings: Option<Vec<String, Global>>,

}

Fields
id: Option<String>

Unique identifier of the daemon.

> Note: The format of the ID itself is not part of the API, and > should not be considered stable.

containers: Option<i64>

Total number of containers on the host.

containers_running: Option<i64>

Number of containers with status \"running\".

containers_paused: Option<i64>

Number of containers with status \"paused\".

containers_stopped: Option<i64>

Number of containers with status \"stopped\".

images: Option<i64>

Total number of images on the host. Both tagged and untagged (dangling) images are counted.

driver: Option<String>

Name of the storage driver in use.

driver_status: Option<Vec<Vec<String, Global>, Global>>

Information specific to the storage driver, provided as "label" / "value" pairs. This information is provided by the storage driver, and formatted in a way consistent with the output of docker info on the command line.

> Note: The information returned in this field, including the > formatting of values and labels, should not be considered stable, > and may change without notice.

docker_root_dir: Option<String>

Root directory of persistent Docker state. Defaults to /var/lib/docker on Linux, and C:\\ProgramData\\docker on Windows.

plugins: Option<PluginsInfo>memory_limit: Option<bool>

Indicates if the host has memory limit support enabled.

swap_limit: Option<bool>

Indicates if the host has memory swap limit support enabled.

kernel_memory: Option<bool>

Indicates if the host has kernel memory limit support enabled.

> Deprecated: This field is deprecated as the kernel 5.4 deprecated > kmem.limit_in_bytes.

kernel_memory_tcp: Option<bool>

Indicates if the host has kernel memory TCP limit support enabled. Kernel memory TCP limits are not supported when using cgroups v2, which does not support the corresponding memory.kmem.tcp.limit_in_bytes cgroup.

cpu_cfs_period: Option<bool>

Indicates if CPU CFS(Completely Fair Scheduler) period is supported by the host.

cpu_cfs_quota: Option<bool>

Indicates if CPU CFS(Completely Fair Scheduler) quota is supported by the host.

cpu_shares: Option<bool>

Indicates if CPU Shares limiting is supported by the host.

cpu_set: Option<bool>

Indicates if CPUsets (cpuset.cpus, cpuset.mems) are supported by the host. See cpuset(7)

pids_limit: Option<bool>

Indicates if the host kernel has PID limit support enabled.

oom_kill_disable: Option<bool>

Indicates if OOM killer disable is supported on the host.

ipv4_forwarding: Option<bool>

Indicates IPv4 forwarding is enabled.

bridge_nf_iptables: Option<bool>

Indicates if bridge-nf-call-iptables is available on the host.

bridge_nf_ip6tables: Option<bool>

Indicates if bridge-nf-call-ip6tables is available on the host.

debug: Option<bool>

Indicates if the daemon is running in debug-mode / with debug-level logging enabled.

n_fd: Option<i64>

The total number of file Descriptors in use by the daemon process. This information is only returned if debug-mode is enabled.

n_goroutines: Option<i64>

The number of goroutines that currently exist. This information is only returned if debug-mode is enabled.

system_time: Option<String>

Current system-time in RFC 3339 format with nano-seconds.

logging_driver: Option<String>

The logging driver to use as a default for new containers.

cgroup_driver: Option<SystemInfoCgroupDriverEnum>

The driver to use for managing cgroups.

cgroup_version: Option<SystemInfoCgroupVersionEnum>

The version of the cgroup.

n_events_listener: Option<i64>

Number of event listeners subscribed.

kernel_version: Option<String>

Kernel version of the host. On Linux, this information obtained from uname. On Windows this information is queried from the HKEY_LOCAL_MACHINE\\SOFTWARE\\Microsoft\\Windows NT\\CurrentVersion\\ registry value, for example "10.0 14393 (14393.1198.amd64fre.rs1_release_sec.170427-1353)".

operating_system: Option<String>

Name of the host’s operating system, for example: "Ubuntu 16.04.2 LTS" or "Windows Server 2016 Datacenter"

os_version: Option<String>

Version of the host’s operating system

> Note: The information returned in this field, including its > very existence, and the formatting of values, should not be considered > stable, and may change without notice.

os_type: Option<String>

Generic type of the operating system of the host, as returned by the Go runtime (GOOS). Currently returned values are "linux" and "windows". A full list of possible values can be found in the Go documentation.

architecture: Option<String>

Hardware architecture of the host, as returned by the Go runtime (GOARCH). A full list of possible values can be found in the Go documentation.

ncpu: Option<i64>

The number of logical CPUs usable by the daemon. The number of available CPUs is checked by querying the operating system when the daemon starts. Changes to operating system CPU allocation after the daemon is started are not reflected.

mem_total: Option<i64>

Total amount of physical memory available on the host, in bytes.

index_server_address: Option<String>

Address / URL of the index server that is used for image search, and as a default for user authentication for Docker Hub and Docker Cloud.

registry_config: Option<RegistryServiceConfig>generic_resources: Option<GenericResourcesInner>http_proxy: Option<String>

HTTP-proxy configured for the daemon. This value is obtained from the HTTP_PROXY environment variable. Credentials (user info component) in the proxy URL are masked in the API response. Containers do not automatically inherit this configuration.

https_proxy: Option<String>

HTTPS-proxy configured for the daemon. This value is obtained from the HTTPS_PROXY environment variable. Credentials (user info component) in the proxy URL are masked in the API response. Containers do not automatically inherit this configuration.

no_proxy: Option<String>

Comma-separated list of domain extensions for which no proxy should be used. This value is obtained from the NO_PROXY environment variable. Containers do not automatically inherit this configuration.

name: Option<String>

Hostname of the host.

labels: Option<Vec<String, Global>>

User-defined labels (key/value metadata) as set on the daemon.

> Note: When part of a Swarm, nodes can both have daemon labels, > set through the daemon configuration, and node labels, set from a > manager node in the Swarm. Node labels are not included in this > field. Node labels can be retrieved using the /nodes/(id) endpoint > on a manager node in the Swarm.

experimental_build: Option<bool>

Indicates if experimental features are enabled on the daemon.

server_version: Option<String>

Version string of the daemon. > Note: the standalone Swarm API > returns the Swarm version instead of the daemon version, for example > swarm/1.2.8.

cluster_store: Option<String>

URL of the distributed storage backend. The storage backend is used for multihost networking (to store network and endpoint information) and by the node discovery mechanism.

> Deprecated: This field is only propagated when using standalone Swarm > mode, and overlay networking using an external k/v store. Overlay > networks with Swarm mode enabled use the built-in raft store, and > this field will be empty.

cluster_advertise: Option<String>

The network endpoint that the Engine advertises for the purpose of node discovery. ClusterAdvertise is a host:port combination on which the daemon is reachable by other hosts.

> Deprecated: This field is only propagated when using standalone Swarm > mode, and overlay networking using an external k/v store. Overlay > networks with Swarm mode enabled use the built-in raft store, and > this field will be empty.

runtimes: Option<HashMap<String, Runtime, RandomState>>

List of OCI compliant runtimes configured on the daemon. Keys hold the "name" used to reference the runtime. The Docker daemon relies on an OCI compliant runtime (invoked via the containerd daemon) as its interface to the Linux kernel namespaces, cgroups, and SELinux. The default runtime is runc, and automatically configured. Additional runtimes can be configured by the user and will be listed here.

default_runtime: Option<String>

Name of the default OCI runtime that is used when starting containers. The default can be overridden per-container at create time.

swarm: Option<SwarmInfo>live_restore_enabled: Option<bool>

Indicates if live restore is enabled. If enabled, containers are kept running when the daemon is shutdown or upon daemon start if running containers are detected.

isolation: Option<SystemInfoIsolationEnum>

Represents the isolation technology to use as a default for containers. The supported values are platform-specific. If no isolation value is specified on daemon start, on Windows client, the default is hyperv, and on Windows server, the default is process. This option is currently not used on other platforms.

init_binary: Option<String>

Name and, optional, path of the docker-init binary. If the path is omitted, the daemon searches the host’s $PATH for the binary and uses the first result.

containerd_commit: Option<Commit>runc_commit: Option<Commit>init_commit: Option<Commit>security_options: Option<Vec<String, Global>>

List of security features that are enabled on the daemon, such as apparmor, seccomp, SELinux, user-namespaces (userns), and rootless. Additional configuration options for each security feature may be present, and are included as a comma-separated list of key/value pairs.

product_license: Option<String>

Reports a summary of the product license on the daemon. If a commercial license has been applied to the daemon, information such as number of nodes, and expiration are included.

default_address_pools: Option<Vec<SystemInfoDefaultAddressPools, Global>>

List of custom default address pools for local networks, which can be specified in the daemon.json file or dockerd option. Example: a Base "10.10.0.0/16" with Size 24 will define the set of 256 10.10.[0-255].0/24 address pools.

warnings: Option<Vec<String, Global>>

List of warnings / informational messages about missing features, or issues related to the daemon configuration. These messages can be printed by the client as information to the user.