Struct bollard::models::TaskSpecContainerSpec

pub struct TaskSpecContainerSpec {

    pub image: Option<String>,
    pub labels: Option<HashMap<String, String, RandomState>>,
    pub command: Option<Vec<String, Global>>,
    pub args: Option<Vec<String, Global>>,
    pub hostname: Option<String>,
    pub env: Option<Vec<String, Global>>,
    pub dir: Option<String>,
    pub user: Option<String>,
    pub groups: Option<Vec<String, Global>>,
    pub privileges: Option<TaskSpecContainerSpecPrivileges>,
    pub tty: Option<bool>,
    pub open_stdin: Option<bool>,
    pub read_only: Option<bool>,
    pub mounts: Option<Vec<Mount, Global>>,
    pub stop_signal: Option<String>,
    pub stop_grace_period: Option<i64>,
    pub health_check: Option<HealthConfig>,
    pub hosts: Option<Vec<String, Global>>,
    pub dns_config: Option<TaskSpecContainerSpecDnsConfig>,
    pub secrets: Option<Vec<TaskSpecContainerSpecSecrets, Global>>,
    pub configs: Option<Vec<TaskSpecContainerSpecConfigs, Global>>,
    pub isolation: Option<TaskSpecContainerSpecIsolationEnum>,
    pub init: Option<bool>,
    pub sysctls: Option<HashMap<String, String, RandomState>>,
    pub capability_add: Option<Vec<String, Global>>,
    pub capability_drop: Option<Vec<String, Global>>,
    pub ulimits: Option<Vec<ResourcesUlimits, Global>>,

}

Container spec for the service.

> Note: ContainerSpec, NetworkAttachmentSpec, and PluginSpec are > mutually exclusive. PluginSpec is only used when the Runtime field > is set to plugin. NetworkAttachmentSpec is used when the Runtime > field is set to attachment.

Fields
image: Option<String>

The image name to use for the container

labels: Option<HashMap<String, String, RandomState>>

User-defined key/value data.

command: Option<Vec<String, Global>>

The command to be run in the image.

args: Option<Vec<String, Global>>

Arguments to the command.

hostname: Option<String>

The hostname to use for the container, as a valid RFC 1123 hostname.

env: Option<Vec<String, Global>>

A list of environment variables in the form VAR=value.

dir: Option<String>

The working directory for commands to run in.

user: Option<String>

The user inside the container.

groups: Option<Vec<String, Global>>

A list of additional groups that the container process will run as.

privileges: Option<TaskSpecContainerSpecPrivileges>tty: Option<bool>

Whether a pseudo-TTY should be allocated.

open_stdin: Option<bool>

Open stdin

read_only: Option<bool>

Mount the container’s root filesystem as read only.

mounts: Option<Vec<Mount, Global>>

Specification for mounts to be added to containers created as part of the service.

stop_signal: Option<String>

Signal to stop the container.

stop_grace_period: Option<i64>

Amount of time to wait for the container to terminate before forcefully killing it.

health_check: Option<HealthConfig>hosts: Option<Vec<String, Global>>

A list of hostname/IP mappings to add to the container’s hosts file. The format of extra hosts is specified in the hosts(5) man page: IP_address canonical_hostname [aliases…]

dns_config: Option<TaskSpecContainerSpecDnsConfig>secrets: Option<Vec<TaskSpecContainerSpecSecrets, Global>>

Secrets contains references to zero or more secrets that will be exposed to the service.

configs: Option<Vec<TaskSpecContainerSpecConfigs, Global>>

Configs contains references to zero or more configs that will be exposed to the service.

isolation: Option<TaskSpecContainerSpecIsolationEnum>

Isolation technology of the containers running the service. (Windows only)

init: Option<bool>

Run an init inside the container that forwards signals and reaps processes. This field is omitted if empty, and the default (as configured on the daemon) is used.

sysctls: Option<HashMap<String, String, RandomState>>

Set kernel namedspaced parameters (sysctls) in the container. The Sysctls option on services accepts the same sysctls as the are supported on containers. Note that while the same sysctls are supported, no guarantees or checks are made about their suitability for a clustered environment, and it’s up to the user to determine whether a given sysctl will work properly in a Service.

capability_add: Option<Vec<String, Global>>

A list of kernel capabilities to add to the default set for the container.

capability_drop: Option<Vec<String, Global>>

A list of kernel capabilities to drop from the default set for the container.

ulimits: Option<Vec<ResourcesUlimits, Global>>

A list of resource limits to set in the container. For example: {\"Name\": \"nofile\", \"Soft\": 1024, \"Hard\": 2048}"